move db-specific methods to parent class
