+++ geoparser/geonames.py
@@ -1,155 +1,12 @@
-import math
-import os
 import re
-import shutil
-import sqlite3
-import typing as t
-import zipfile
 
-import pandas as pd
-import requests
-from tqdm.auto import tqdm
-
-from geoparser.config import get_gazetteer_configs
-from geoparser.config.models import GazetteerData
-from geoparser.gazetteer import Gazetteer
+from geoparser.gazetteer import LocalDBGazetteer
 
 
-class GeoNames(Gazetteer):
+class GeoNames(LocalDBGazetteer):
     def __init__(self):
         super().__init__("geonames")
-        self.config = get_gazetteer_configs()["geonames"]
         self.location_description_template = "<name> (<feature_name>) COND[in, any{<admin2_name>, <admin1_name>, <country_name>}] <admin2_name>, <admin1_name>, <country_name>"
-
-    def setup_database(self):
-        print("Database setup...")
-
-        self.clean_dir()
-
-        for dataset in self.config.data:
-            self.download_data(dataset.url)
-            self.load_data(dataset)
-
-        self.clean_dir(keep_db=True)
-
-        print("Database setup complete.")
-
-    def clean_dir(self, keep_db: bool = False):
-
-        if os.path.exists(self.data_dir):
-            for file_name in os.listdir(self.data_dir):
-                if keep_db and (
-                    file_name.endswith(".db") or file_name.endswith(".db-journal")
-                ):
-                    continue
-                else:
-                    try:
-                        os.remove(os.path.join(self.data_dir, file_name))
-                    except IsADirectoryError:
-                        os.rmdir(os.path.join(self.data_dir, file_name))
-                    except PermissionError:
-                        shutil.rmtree(os.path.join(self.data_dir, file_name))
-
-    def download_data(self, url: str):
-        os.makedirs(self.data_dir, exist_ok=True)
-        self.download_file(url)
-
-    def download_file(self, url: str):
-        filename = url.split("/")[-1]
-        file_path = os.path.join(self.data_dir, filename)
-        response = requests.get(url, stream=True)
-        response.raise_for_status()
-        with open(file_path, "wb") as f, tqdm(
-            desc=f"Downloading {filename}",
-            total=int(response.headers.get("content-length", 0)),
-            unit="iB",
-            unit_scale=True,
-            unit_divisor=1024,
-        ) as bar:
-            for chunk in response.iter_content(chunk_size=1024):
-                size = f.write(chunk)
-                bar.update(size)
-        if file_path.endswith(".zip"):
-            with zipfile.ZipFile(file_path, "r") as zip_ref:
-                zip_ref.extractall(self.data_dir)
-            os.remove(file_path)
-
-    def load_data(self, dataset: GazetteerData):
-
-        conn = sqlite3.connect(self.db_path)
-
-        self.create_tables(conn, dataset)
-        self.populate_tables(conn, dataset)
-
-        conn.close()
-
-    def create_tables(self, conn: sqlite3.Connection, dataset: GazetteerData):
-
-        cursor = conn.cursor()
-
-        columns = ", ".join(
-            [
-                f"{col.name} {col.type}{' PRIMARY KEY' if col.primary else ''}"
-                for col in dataset.columns
-            ]
-        )
-        cursor.execute(f"CREATE TABLE IF NOT EXISTS {dataset.name} ({columns})")
-
-        for virtual_table in dataset.virtual_tables:
-            args = ", ".join(virtual_table.args)
-            kwargs = ", ".join(
-                [f'{key}="{value}"' for key, value in virtual_table.kwargs.items()]
-            )
-            cursor.execute(
-                f"CREATE VIRTUAL TABLE IF NOT EXISTS {virtual_table.name} USING {virtual_table.using}({args}, {kwargs})"
-            )
-
-        conn.commit()
-
-    def populate_tables(self, conn: sqlite3.Connection, dataset: GazetteerData):
-
-        cursor = conn.cursor()
-        self.load_file_into_table(
-            conn,
-            os.path.join(self.data_dir, dataset.extracted_file),
-            dataset.name,
-            [col.name for col in dataset.columns],
-            skiprows=dataset.skiprows,
-        )
-
-        for virtual_table in dataset.virtual_tables:
-            cursor.execute(
-                f"INSERT INTO {virtual_table.name} (rowid, {virtual_table.args[0]}) SELECT {virtual_table.kwargs['content_rowid']}, {virtual_table.args[0]} FROM {dataset.name}"
-            )
-
-        conn.commit()
-
-    def load_file_into_table(
-        self,
-        conn: sqlite3.Connection,
-        file_path: str,
-        table_name: str,
-        columns: list[str],
-        skiprows: t.Union[int, list[int], t.Callable] = None,
-        chunksize: int = 100000,
-    ):
-
-        chunks = pd.read_csv(
-            file_path,
-            delimiter="\t",
-            header=None,
-            names=columns,
-            chunksize=chunksize,
-            dtype=str,
-            skiprows=skiprows,
-        )
-
-        for chunk in tqdm(
-            chunks,
-            desc=f"Loading {table_name}",
-            total=math.ceil(sum(1 for row in open(file_path, "rb")) / chunksize),
-        ):
-            chunk.to_sql(table_name, conn, if_exists="append", index=False)
 
     def query_candidates(
         self,
